Pulse length is now calculated as len = MIN(s, pulseLength), where s is the distance the pulse has moved since it was shot or bounced the last time. It is assumed the pulse moves at a constant speed (pulseSpeed).

+++ xpilot/src/server/frame.c
@@ -953,7 +953,7 @@ static void Frame_shots(int conn, int ind)
 		} else {
 		    color = RED;
 		}
-		Send_laser(conn, color, lcx, lcy, pulse->len, ldir);
+		Send_laser(conn, color, lcx, lcy, (int)pulse->len, ldir);
 	    }
 	break;
 	default:
